requirements.txt
-i https://pypi.org/simple
aiohttp==3.8.4 ; python_version >= '3.6'
aiosignal==1.3.1 ; python_version >= '3.7'
alembic==1.11.1 ; python_version >= '3.7'
anyio==3.7.0 ; python_version >= '3.7'
apache-airflow==2.6.1
apache-airflow-providers-amazon==8.1.0
apache-airflow-providers-common-sql==1.5.1 ; python_version ~= '3.7'
apache-airflow-providers-ftp==3.4.1 ; python_version ~= '3.7'
apache-airflow-providers-http==4.4.1
apache-airflow-providers-imap==3.2.1 ; python_version ~= '3.7'
apache-airflow-providers-sqlite==3.4.1 ; python_version ~= '3.7'
apispec[yaml]==5.2.2 ; python_version >= '3.7'
argcomplete==3.0.8 ; python_version >= '3.6'
asgiref==3.7.2 ; python_version >= '3.7'
asn1crypto==1.5.1
async-timeout==4.0.2 ; python_version >= '3.6'
attrs==23.1.0 ; python_version >= '3.7'
babel==2.12.1 ; python_version >= '3.7'
beautifulsoup4==4.12.2 ; python_full_version >= '3.6.0'
blinker==1.6.2 ; python_version >= '3.7'
boto3==1.26.146 ; python_version >= '3.7'
botocore==1.29.146 ; python_version >= '3.7'
cachelib==0.9.0 ; python_version >= '3.7'
cattrs==23.1.2 ; python_version >= '3.7'
certifi==2023.5.7 ; python_version >= '3.6'
cffi==1.15.1
charset-normalizer==3.1.0 ; python_full_version >= '3.7.0'
click==8.1.3 ; python_version >= '3.7'
clickclick==20.10.2
colorama==0.4.6 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
colorlog==4.8.0
configupdater==3.1.1 ; python_version >= '3.6'
connexion[flask]==2.14.2 ; python_version >= '3.6'
cron-descriptor==1.4.0
croniter==1.3.15 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
cryptography==41.0.1 ; python_version >= '3.7'
decorator==5.1.1 ; python_version >= '3.5'
deprecated==1.2.14 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
dill==0.3.6 ; python_version >= '3.7'
dnspython==2.3.0 ; python_version >= '3.7' and python_version < '4.0'
docutils==0.20.1 ; python_version >= '3.7'
email-validator==1.3.1 ; python_version >= '3.5'
exceptiongroup==1.1.1 ; python_version < '3.11'
flask==2.2.5 ; python_version >= '3.7'
flask-appbuilder==4.3.0 ; python_version ~= '3.7'
flask-babel==2.0.0
flask-caching==2.0.2 ; python_version >= '3.7'
flask-jwt-extended==4.5.2 ; python_version >= '3.7' and python_version < '4'
flask-limiter==3.3.1 ; python_version >= '3.7'
flask-login==0.6.2 ; python_version >= '3.7'
flask-session==0.5.0 ; python_version >= '3.7'
flask-sqlalchemy==2.5.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
flask-wtf==1.1.1 ; python_version >= '3.7'
frozenlist==1.3.3 ; python_version >= '3.7'
graphviz==0.20.1 ; python_version >= '3.7'
greenlet==2.0.2 ; python_version >= '3' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
gunicorn==20.1.0 ; python_version >= '3.5'
h11==0.14.0 ; python_version >= '3.7'
httpcore==0.17.2 ; python_version >= '3.7'
httpx==0.24.1 ; python_version >= '3.7'
idna==3.4 ; python_version >= '3.5'
importlib-resources==5.12.0 ; python_version >= '3.7'
inflection==0.5.1 ; python_version >= '3.5'
itsdangerous==2.1.2 ; python_version >= '3.7'
jinja2==3.1.2 ; python_version >= '3.7'
jmespath==1.0.1 ; python_version >= '3.7'
jsonpath-ng==1.5.3
jsonschema==4.17.3 ; python_version >= '3.7'
lazy-object-proxy==1.9.0 ; python_version >= '3.7'
limits==3.5.0 ; python_version >= '3.7'
linkify-it-py==2.0.2 ; python_version >= '3.7'
lockfile==0.12.2
lxml==4.9.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
mako==1.2.4 ; python_version >= '3.7'
markdown==3.4.3 ; python_version >= '3.7'
markdown-it-py==2.2.0 ; python_version >= '3.7'
markupsafe==2.1.3 ; python_version >= '3.7'
marshmallow==3.19.0 ; python_version >= '3.7'
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1 ; python_version >= '3.6'
marshmallow-sqlalchemy==0.26.1 ; python_version >= '3.6'
mdit-py-plugins==0.3.5 ; python_version >= '3.7'
mdurl==0.1.2 ; python_version >= '3.7'
multidict==6.0.4 ; python_version >= '3.7'
mypy-boto3-appflow==1.26.145 ; python_version >= '3.7'
mypy-boto3-rds==1.26.144 ; python_version >= '3.7'
mypy-boto3-redshift-data==1.26.109 ; python_version >= '3.7'
mypy-boto3-s3==1.26.127 ; python_version >= '3.7'
ordered-set==4.1.0 ; python_version >= '3.7'
packaging==23.1 ; python_version >= '3.7'
pathspec==0.9.0
pendulum==2.1.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pluggy==1.0.0 ; python_version >= '3.6'
ply==3.11
prison==0.2.1
psutil==5.9.5 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pycparser==2.21
pydantic==1.10.8 ; python_version >= '3.7'
pygments==2.15.1 ; python_version >= '3.7'
pyjwt==2.7.0 ; python_version >= '3.7'
pyrsistent==0.19.3 ; python_version >= '3.7'
python-daemon==3.0.1 ; python_version >= '3'
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-nvd3==0.15.0
python-slugify==8.0.1 ; python_version >= '3.7'
pytz==2023.3
pytzdata==2020.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyyaml==6.0 ; python_version >= '3.6'
redshift-connector==2.0.910 ; python_version >= '3.6'
requests==2.31.0 ; python_version >= '3.7'
requests-toolbelt==1.0.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rfc3339-validator==0.1.4 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
rich==13.4.1 ; python_full_version >= '3.7.0'
rich-argparse==1.1.1 ; python_version >= '3.7'
s3transfer==0.6.1 ; python_version >= '3.7'
scramp==1.4.4 ; python_version >= '3.7'
setproctitle==1.3.2 ; python_version >= '3.7'
setuptools==67.8.0 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0 ; python_version >= '3.7'
soupsieve==2.4.1 ; python_version >= '3.7'
sqlalchemy==1.4.48 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
sqlalchemy-jsonfield==1.0.1.post0 ; python_full_version >= '3.7.0'
sqlalchemy-redshift==0.8.14 ; python_version >= '3.4'
sqlalchemy-utils==0.41.1 ; python_version >= '3.6'
sqlparse==0.4.4 ; python_version >= '3.5'
tabulate==0.9.0 ; python_version >= '3.7'
tenacity==8.2.2 ; python_version >= '3.6'
termcolor==2.3.0 ; python_version >= '3.7'
text-unidecode==1.3
typing-extensions==4.6.3 ; python_version >= '3.7'
uc-micro-py==1.0.2 ; python_version >= '3.7'
unicodecsv==0.14.1
urllib3==1.26.16 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
watchtower==2.0.1 ; python_version >= '3.6'
werkzeug==2.2.3 ; python_version >= '3.7'
wrapt==1.15.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
wtforms==3.0.1 ; python_version >= '3.7'
yarl==1.9.2 ; python_version >= '3.7'
